requirement.txt
openai==0.28
aiogram
python-dotenv